Ratings recyclkerview + link - mancano i layout

+++ app/src/main/java/com/example/myapplication/data/Review.kt
@@ -2,7 +2,7 @@ package com.example.myapplication.data
 
 import java.io.Serializable
 
-data class Review(val reviewedBy : String,val reviewedByUsername : String, val stars : Float, val comment: String? = null) :
+data class Review(val reviewedBy : String, val reviewedByUsername : String, val stars : Double, val comment: String? = null) :
     Serializable {
 
     companion object RatingFactory{
@@ -10,7 +10,7 @@ data class Review(val reviewedBy : String,val reviewedByUsername : String, val s
         fun fromMapToObj(hash : Map<String, Any>?) : Review {
             val reviewedBy = hash?.get("reviewedBy") as String
             val reviewedByUsername = hash["reviewedByUsername"] as String
-            val rating = hash["stars"] as Float
+            val rating = hash["stars"] as Double
             val comment = hash["comment"] as String?
 
             return Review(reviewedBy, reviewedByUsername, rating, comment)
